moves v3 transformers to root level
