+++ app/Http/Controllers/TagsController.php
@@ -6,7 +6,7 @@ use Rogue\Models\Post;
 use Illuminate\Http\Request;
 use Rogue\Repositories\Three\PostRepository;
 use Rogue\Http\Controllers\Legacy\Two\ApiController;
-use Rogue\Http\Transformers\Three\PostTransformer;
+use Rogue\Http\Transformers\PostTransformer;
 
 class TagsController extends ApiController
 {
@@ -18,7 +18,7 @@ class TagsController extends ApiController
     protected $post;
 
     /**
-     * @var \League\Fractal\TransformerAbstract;
+     * @var Rogue\Http\Transformers\PostTransformer;
      */
     protected $transformer;
