+++ app/Http/Transformers/SignupTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Transformers\Three;
+namespace Rogue\Http\Transformers;
 
 use Rogue\Models\Signup;
 use Rogue\Services\Registrar;
